Operation Exclusion Changes

+++ restservices/src/main/java/com/osi/urm/service/dto/OsiFunctionsDTO.java
@@ -11,7 +11,7 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 @JsonInclude(Include.NON_EMPTY)
 public class OsiFunctionsDTO implements java.io.Serializable {
 
-	private Integer id;
+	private Long id;
 	private String funcName;
 	private String funcType;
 	private String funcValue;
@@ -30,14 +30,14 @@ public class OsiFunctionsDTO implements java.io.Serializable {
 	public OsiFunctionsDTO() {
 	}
 
-	public OsiFunctionsDTO(Integer id, String funcType, String funcValue,String funcName) {
+	public OsiFunctionsDTO(Long id, String funcType, String funcValue,String funcName) {
 		this.id = id;
 		this.funcName = funcName;
 		this.funcType = funcType;
 		this.funcValue = funcValue;
 	}
 
-	public OsiFunctionsDTO(Integer id, String funcType, String funcValue,
+	public OsiFunctionsDTO(Long id, String funcType, String funcValue,
 			String parameters, Integer createdBy, Date createdDate,String funcName,
 			Integer updatedBy, Date updatedDate,
 			Set<OsiUserFuncExclDTO> osiUserFuncExcls,
@@ -57,11 +57,11 @@ public class OsiFunctionsDTO implements java.io.Serializable {
 		this.osiUserOperationExcls = osiUserOperationExcls;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
